transfer starting point to deku component
