+++ lib/index.js
@@ -1,9 +1,19 @@
+/* eslint-disable deku/no-unknown-property */
+
 import {render, tree} from 'deku';
 import element from 'magic-virtual-element';
 import {save as saveSelection, restore as restoreSelection} from 'save-selection';
 import morphdom from 'morphdom';
 import setupHtmlToArticleJson from 'html-to-article-json';
-import _Component from './component';
+import assert from 'assert';
+import setupArticle from 'article-json-html-render';
+import embeds from './embeds';
+import setupEvents from './setup-events';
+
+const Article = setupArticle({
+  embeds: embeds,
+  renderEmptyTextNodes: true
+});
 
 const htmlToArticleJson = setupHtmlToArticleJson({
   customEmbedTypes: [{
@@ -18,28 +28,32 @@ const htmlToArticleJson = setupHtmlToArticleJson({
   }]
 });
 
-export const Component = _Component;
+export default {
+  name: 'Wrapper',
+  render: ({props: {onInput}}) => {
+    const _onInput = ({delegateTarget}) => {
+      saveSelection(delegateTarget);
+      const newArticleJson = htmlToArticleJson(delegateTarget);
+      onInput({ items: newArticleJson });
+    };
 
-export const getComponent = ({items, onInput}) => {
-  const update = setupUpdate({onInput});
-  const _onInput = ({delegateTarget}) => {
-    update({delegateTarget});
-  };
+    const events = setupEvents(_onInput);
+    return (<div contenteditable='true' {...events}>
+      <Article items={[]} />
+    </div>);
+  },
+  afterRender: ({props: {items}}, elm) => {
+    const oldArticleElm = elm.querySelector('article');
+    const app = tree(<Article items={items} />);
 
-  return (<Component items={items} onInput={_onInput} />);
-};
-
-function setupUpdate ({onInput}) {
-  return ({delegateTarget}) => {
-    saveSelection(delegateTarget);
-    const newArticleJson = htmlToArticleJson(delegateTarget);
-    onInput({ items: newArticleJson });
     const tmpElm = document.createElement('div');
-    const app = tree(getComponent({items: newArticleJson, onInput}));
     render(app, tmpElm);
-    const oldArticleElm = delegateTarget.querySelector('article');
     const newArticleElm = tmpElm.querySelector('article');
+
+    assert(oldArticleElm, 'oldArticleElm must exists');
+    assert(newArticleElm, 'newArticleElm must exists');
+
     morphdom(oldArticleElm, newArticleElm);
-    restoreSelection(delegateTarget);
-  };
-}
+    restoreSelection(oldArticleElm);
+  }
+};
